### 2023-04-24 #5401 @oliviaguest
Hi @jmotis, thank you for choosing JOSS. Can you try to address the missing stuff the bot has noticed above, please? After that, feel free to recompile the PDF to check what everything looks like. Thank you! 😊
