### 2023-06-21 #5401 @arfon
@jmotis hello! I'm going to be the editor handling your paper. Could you take a look a [our reviewers tool which has a list of potential reviewers](https://reviewers.joss.theoj.org/) and identify a few people who would be good candidates to review this submission?
