### 2023-07-01 #5401 @arfon
:wave:  @ikayz @chrisbrickhouse @luxaritas – would any of you be willing to review this submission for JOSS? The submission under consideration is DataScribe: An Omeka S module for structured data transcription.

Many thanks
Arfon
