### 2023-07-16 #5401 @arfon
:wave: @koenedaele – you be willing to review this submission for JOSS? The submission under consideration is DataScribe: An Omeka S module for structured data transcription.

Many thanks
Arfon
