### 2023-07-17 #5661 @editorialbot
[REVIEW]: DataScribe: An Omeka S module for structured data transcription
**Submitting author:** @jmotis (<a href="http://orcid.org/0000-0001-5519-8331">Jessica Otis</a>)
**Repository:** https://github.com/chnm/Datascribe-module
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v1.1.0
