### 2023-07-26 #5661 @koenedaele
Hi @jmotis, I'm one of the reviewers having a look at your software. I'll conduct the review by creating issues in your repository to keep this thread clean as requested by JOSS.
