### 2023-07-26 chnm/Datascribe-module#131 @koenedaele
[JOSS] Please use correct style of citation
Part of openjournals/joss-reviews#5661

The paper does not seem to adhere to the required citation style. The citations are not in a bibtex file and the citations do not follow Pandoc format. Among other things, URL's are not hyperlinked and the bibliography generally looks a bit off.
Perhaps have a look the the example JOSS paper and bibliography (https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) to get started.
