### 2023-08-07 chnm/Datascribe-module#132 @koenedaele
[JOSS] Consider adding a CONTRIBUTING.md file
Part of openjournals/joss-reviews#5661

It's best practice for FOSS to include a CONTRIBUTING.md file, that describes (briefly) how to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. It does not need to contain everything, but some major warnings and links to other parts of your documentation to make sure people pick up on your guidelines. A major advantage of a CONTRIBUTING.md file is that it will be shown to everybody sending a PR, leading to better PR's.

Some docs can be found at
https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors or https://mozillascience.github.io/working-open-workshop/contributing/
