### 2023-08-07 chnm/Datascribe-module#133 @koenedaele
[JOSS] Consider having automated tests
Part of openjournals/joss-reviews#5661

I haven't found any automated tests (PHPUnit, Codeception, Behat, ...) or CI (Travis, Github Actions). While not strictly neccessary, they offer an assurance that things are working as expected, make it harder for new developers to break stuff and also serve as a form of documentation. Have you considered adding them?
