### 2023-08-07 #5661 @koenedaele
I have reviewed the Software Paper and started on some of the documentation. I will be away for a bit. When I get back I'll have a further look at the documentation and the software itself.
