### 2023-08-27 #5661 @luxaritas
Hi all, sorry for the delay on my end! Starting to work my way through this, and Koen has already filed some good issues.

@jmotis Since you do not appear to be a frequent committer to the codebase, can you describe your role as a contributor to the project? Additionally, I note Kim Nguyen as one of the larger committers to the project but is not a listed author, and I want to verify whether lack of authorship there is appropriate.
