### 2023-09-05 #5661 @jmotis
This was a grant-funded project. I was the PI (along with Lincoln Mullen as co-PI) and did a lot of different things: project management and reporting, developing the required feature set, consulting and testing throughout the development of the module, documentation, and outreach. Kim was invited to be on the paper but didn't join us.
