### 2023-09-18 chnm/Datascribe-module#134 @luxaritas
[JOSS] Consider adjusting documentation screenshots with more concrete values
Part of https://github.com/openjournals/joss-reviews/issues/5661

Bit of a nitpick on the really nice docs, open to opinions on how useful or not this would be. You might want to consider tweaking the screenshots in the documentation to be a bit more concrete. It seems like a lot of screenshots have completely empty form fields, where having example data filled in could help clarify the usage and help strengthen the ask for "solving real-world analysis problems".
