### 2023-09-19 chnm/Datascribe-module#134 @jmotis
Thanks for the suggestion. We actually have built up a whole series of case studies and tutorial videos that use real-world data from our various projects but it's on the [project website](https://datascribe.tech) not the GitHub. The website links here but the GitHub doesn't link back, which is an easy fix.
