### 2023-09-19 43b2e06
add website link to readme
linking back to https://datascribe.tech in the readme to guide users to additional documentation
