+++ README.md
@@ -4,6 +4,6 @@ An [Omeka S](https://omeka.org/s/) module for the transcription of structured da
 
 When cloning this repository remember to rename the directory from "Datascribe-module" to "Datascribe".
 
-A user manual is available [in the wiki for this GitHub repository](https://github.com/chnm/Datascribe-module/wiki). 
+A user manual is available [in the wiki for this GitHub repository](https://github.com/chnm/Datascribe-module/wiki). Additional documentation including case studies, video tutorials, and a quick start guide are available on the project website [https://datascribe.tech](https://datascribe.tech).
 
 Note: in order to be able to export datasets, the directory `/files/asset` in your Omeka S installation must be writable by the system.
