### 2023-09-19 #5661 @jmotis
Thanks both of you for your time and feedback. I've linked the GitHub repository to the website where we have the concrete case studies and tutorials (vs. the abstracted documentation in the repo) and we'll tidy up the bibliography and add the CONTRIBUTING.md file.
