### 2023-09-19 62001db
adding CONTRIBUTING.md file
adding first draft of file describing how to contribute to the repository
