+++ CONTRIBUTING.md
@@ -1,0 +1,43 @@
+## How to contribute to DataScribe
+
+Thanks for your interest in contributing to DataScribe (and to the Omeka community more generally).
+
+#### **Did you find a bug?**
+
+* Please check to see if the bug was already reported by checking the open [Issues](https://github.com/chnm/Datascribe-module/issues).
+
+* If you're unable to find an open issue on the bug, [open a new issue](https://github.com/chnm/Datascribe-module/issues/new). Include a title that briefly describes the issue as well as a longer, clear description of the bug.
+
+* A good bug description is likely to include
+    * what browser version you are using (e.g. Safari Version 16.6)
+    * what page or pages you were on when the error occurred
+    * what, if any, actions you took that resulted in the bug
+    * any error message produced by the system
+
+#### **Did you fix a bug?**
+
+* Please check to see if the bug was already reported by checking the open [Issues](https://github.com/chnm/Datascribe-module/issues).  
+
+* If you're unable to find an open issue on the bug, please [open a new issue](https://github.com/chnm/Datascribe-module/issues/new) as described above.
+
+* If an open issue already exists, go ahead and open a new GitHub pull request with the bug fix.
+
+* The pull request description should clearly describe the problem and solution with a reference to the appropriate open issue.
+
+#### **Do you want to add a new feature?**
+
+* Please contact us at [chnm@gmu.edu](mailto:chnm@gmu.edu) with the subject line: Proposed DataScribe Contribution and you will be forwarded to a member of the team to discuss the feature development.
+
+* Do not open an issue on GitHub until you have talked with a member of the team.
+
+#### **Do you have questions about the code?**
+
+* You can ask general questions in the Omeka community forum. You may also contact us at [chnm@gmu.edu](mailto:chnm@gmu.edu) with the subject line: Question about DataScribe.
+
+* Do not open an issue on GitHub with questions about the code or general tech support requests.
+
+#### **Do you want to contribute to the documentation?**
+
+* If you're interested in contributing to the documentation, please contact us at [chnm@gmu.edu](mailto:chnm@gmu.edu) with the subject line: DataScribe Documentation.
+
+* We would particularly welcome people who are interested in helping us translate our documentation into additional languages.
