### 2023-09-19 chnm/Datascribe-module#132 @jmotis
@lmullen I've added a first draft of this file, let me know if you think we need to make any updates
