### 2023-09-28 chnm/Datascribe-module#135 @koenedaele
[JOSS] Install link not working
Part of openjournals/joss-reviews#5661

At https://datascribe.tech/support/quickstart/, the install points to https://omeka.org/s/modules/Datascribe-module/ but this page does not exist.
