### 2023-09-28 chnm/Datascribe-module#135 @koenedaele
The correct link seems to be https://omeka.org/s/modules/Datascribe/
