### 2023-09-28 chnm/Datascribe-module#136 @koenedaele
[JOSS] Datascribe does not work with Omeka S 4
Part of openjournals/joss-reviews#5661

I followed the install instructions at https://datascribe.tech/support/quickstart/. Because I'm not an Omeka S user, I had to install from scratch. Right now (september 28) this means installing version 4.0.4.

However, Datascribe has been released as compatible with Omeka S ^3.0.0., meaning I couldn't finish the installation. If this is intentional, and Datascribe is not compatible with Omeka S 4, please include this in the installation docs. If this is not intentional, please release a new version of the module with the correct version compatibility. If you do plan to stick with version 3 for the forseeable future, what is your longterm plan?

Please let me know if I need to redo the installation to be able to test Datascribe or if a fix is coming soon.
