### 2023-09-28 chnm/Datascribe-module#136 @jmotis
For now, please do reinstall with v3 and thanks for catching that in the documentation.

We're in talks with various folks about upgrade plans but nothing is concrete at this moment.
