### 2023-09-28 chnm/Datascribe-module#136 @hepplerj
Note about DataScribe's compatibility with Omeka S has been added as of chnm/datascribe-site@745ef1d684499994757658251558fb66972cf005.
