### 2023-10-02 #5661 @koenedaele
I have concluded my review. I have found the software to function as indicated. Being new to Omeka S, I had to learn about both Omeka S and Datascribe. I am especially pleased with the detailed documentation of the  module. Congratulations on the authors, writing documentation is never an easy task.

I have created some minor issues that could be solved. I am aware the module only functions for the previous version of Omeka S, but I assume that's not an issue for JOSS (I have no information indicating it is). I do hope the module is updated to the newer version sooner or later.

Currently I'm awaiting a newer version of the paper with proper references, before I sign off on that last aspect. Once the citations are properly formatted, I'm ok with this paper.
