### 2023-10-02 #5661 @luxaritas
Just did one more pass and agree, references is the only remaining thing to be resolved
