### 2023-12-07 1deb693
adding BibLaTeX bibliography
