+++ paper.bib
@@ -1,0 +1,74 @@
+
+@online{adasme_peste_2022,
+	title = {Peste Bubónica en Iquique, 1903},
+	url = {https://datascribe.tech/casestudies/PesteBubonica_Iquique1903.pdf},
+	titleaddon = {{DataScribe}.tech},
+	author = {Adasme, Hernán},
+	date = {2022},
+}
+
+@online{adasme_plague_2022,
+	title = {Plague in Iquique, 1903},
+	url = {https://datascribe.tech/casestudies/Plague_Iquique1903.pdf},
+	titleaddon = {{DataScribe}.tech},
+	author = {Adasme, Hernán},
+	date = {2022},
+}
+
+@online{adasme_death_2022,
+	title = {Death by Numbers},
+	url = {https://datascribe.tech/casestudies/DataScribe_BillsOfMortality_CaseStudy.pdf},
+	titleaddon = {{DataScribe}.tech},
+	author = {Adasme, Hernán and Howlett, Daniel and Meyers, Emily},
+	date = {2022},
+}
+
+@online{brett_1950_2022,
+	title = {1950 United States Census},
+	url = {https://datascribe.tech/casestudies/DataScribeCaseStudy_1950CensusUS.pdf},
+	titleaddon = {{DataScribe}.tech},
+	author = {Brett, Megan},
+	date = {2022},
+}
+
+@online{swain_religious_2022,
+	title = {Religious Ecologies},
+	url = {https://datascribe.tech/casestudies/CaseStudy_AmericanReligiousEcologies.pdf},
+	titleaddon = {{DataScribe}.tech},
+	author = {Swain, Greta},
+	date = {2022},
+}
+
+@software{noauthor_scripto_nodate,
+	title = {Scripto},
+	url = {https://scripto.org},
+}
+
+@software{noauthor_tesseract_nodate,
+	title = {Tesseract},
+	url = {https://tesseract-ocr.github.io},
+}
+
+@software{noauthor_zooniverse_nodate,
+	title = {Zooniverse Project Builder},
+	url = {https://www.zooniverse.org/lab},
+}
+
+@software{noauthor_abbyy_nodate,
+	title = {{ABBYY} {FineReader}},
+	url = {https://pdf.abbyy.com},
+}
+
+@software{death_by_numbers_project_team_death_2016,
+	title = {Death by Numbers},
+	url = {https://deathbynumbers.org/},
+	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
+	author = {Death by Numbers Project Team}
+}
+
+@software{mapping_religious_ecologies_project_team_mapping_2018,
+	title = {Mapping Religious Ecologies},
+	url = {https://religiousecologies.org},
+	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
+	author = {Mapping Religious Ecologies Project Team}
+}
