### 2023-12-07 826f6c8
update citations in paper
