### 2023-12-07 39679af
removing reference list from main paper
should now compile from paper.bib
