### 2023-12-07 ea83773
replacing bib file with different format
