### 2023-12-07 b42a546
Update paper.md
experimenting probably uselessly
