### 2023-12-07 e092c90
Update paper.md
another attempt to debug why references not compiling to the file correctly
