+++ paper.bib
@@ -1,76 +1,74 @@
 
-@misc{adasme_peste_2022,
-	title = {Peste {Bubónica} en {Iquique}, 1903},
+@online{adasme:2022a,
+	title = {Peste Bubónica en Iquique, 1903},
 	url = {https://datascribe.tech/casestudies/PesteBubonica_Iquique1903.pdf},
-	journal = {DataScribe.tech},
+	titleaddon = {{DataScribe}.tech},
 	author = {Adasme, Hernán},
-	year = {2022},
+	date = {2022},
 }
 
-@misc{adasme_plague_2022,
-	title = {Plague in {Iquique}, 1903},
+@online{adasme:2022b,
+	title = {Plague in Iquique, 1903},
 	url = {https://datascribe.tech/casestudies/Plague_Iquique1903.pdf},
-	journal = {DataScribe.tech},
+	titleaddon = {{DataScribe}.tech},
 	author = {Adasme, Hernán},
-	year = {2022},
+	date = {2022},
 }
 
-@misc{adasme_death_2022,
-	title = {Death by {Numbers}},
+@online{adasme:2022c,
+	title = {Death by Numbers},
 	url = {https://datascribe.tech/casestudies/DataScribe_BillsOfMortality_CaseStudy.pdf},
-	journal = {DataScribe.tech},
+	titleaddon = {{DataScribe}.tech},
 	author = {Adasme, Hernán and Howlett, Daniel and Meyers, Emily},
-	year = {2022},
+	date = {2022},
 }
 
-@misc{brett_1950_2022,
-	title = {1950 {United} {States} {Census}},
+@online{brett:2022,
+	title = {1950 United States Census},
 	url = {https://datascribe.tech/casestudies/DataScribeCaseStudy_1950CensusUS.pdf},
-	journal = {DataScribe.tech},
+	titleaddon = {{DataScribe}.tech},
 	author = {Brett, Megan},
-	year = {2022},
+	date = {2022},
 }
 
-@misc{swain_religious_2022,
-	title = {Religious {Ecologies}},
+@online{swain:2022,
+	title = {Religious Ecologies},
 	url = {https://datascribe.tech/casestudies/CaseStudy_AmericanReligiousEcologies.pdf},
-	journal = {DataScribe.tech},
+	titleaddon = {{DataScribe}.tech},
 	author = {Swain, Greta},
-	year = {2022},
+	date = {2022},
 }
 
-@misc{noauthor_scripto_nodate,
+@software{scripto,
 	title = {Scripto},
 	url = {https://scripto.org},
 }
 
-@misc{noauthor_tesseract_nodate,
+@software{tesseract,
 	title = {Tesseract},
 	url = {https://tesseract-ocr.github.io},
 }
 
-@misc{noauthor_zooniverse_nodate,
-	title = {Zooniverse {Project} {Builder}},
+@software{zooniverse,
+	title = {Zooniverse Project Builder},
 	url = {https://www.zooniverse.org/lab},
 }
 
-@misc{noauthor_abbyy_nodate,
+@software{abbyy,
 	title = {{ABBYY} {FineReader}},
 	url = {https://pdf.abbyy.com},
 }
 
-@misc{death_by_numbers_project_team_death_2016,
-	title = {Death by {Numbers}},
+@software{death,
+	title = {Death by Numbers},
 	url = {https://deathbynumbers.org/},
 	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
-	author = {Death by Numbers Project Team},
-	year = {2016},
+	author = {Death by Numbers Project Team}
 }
 
-@misc{mapping_religious_ecologies_project_team_mapping_2018,
-	title = {Mapping {Religious} {Ecologies}},
+@software{mapping,
[…]
