### 2023-12-07 e1beb66
more manual bib tweaks
