+++ paper.bib
@@ -1,5 +1,5 @@
 
-@online{adasme:2022a,
+@misc{adasme:2022a,
 	title = {Peste Bubónica en Iquique, 1903},
 	url = {https://datascribe.tech/casestudies/PesteBubonica_Iquique1903.pdf},
 	titleaddon = {{DataScribe}.tech},
@@ -7,7 +7,7 @@
 	date = {2022},
 }
 
-@online{adasme:2022b,
+@misc{adasme:2022b,
 	title = {Plague in Iquique, 1903},
 	url = {https://datascribe.tech/casestudies/Plague_Iquique1903.pdf},
 	titleaddon = {{DataScribe}.tech},
@@ -15,7 +15,7 @@
 	date = {2022},
 }
 
-@online{adasme:2022c,
+@misc{adasme:2022c,
 	title = {Death by Numbers},
 	url = {https://datascribe.tech/casestudies/DataScribe_BillsOfMortality_CaseStudy.pdf},
 	titleaddon = {{DataScribe}.tech},
@@ -23,7 +23,7 @@
 	date = {2022},
 }
 
-@online{brett:2022,
+@misc{brett:2022,
 	title = {1950 United States Census},
 	url = {https://datascribe.tech/casestudies/DataScribeCaseStudy_1950CensusUS.pdf},
 	titleaddon = {{DataScribe}.tech},
@@ -31,7 +31,7 @@
 	date = {2022},
 }
 
-@online{swain:2022,
+@misc{swain:2022,
 	title = {Religious Ecologies},
 	url = {https://datascribe.tech/casestudies/CaseStudy_AmericanReligiousEcologies.pdf},
 	titleaddon = {{DataScribe}.tech},
@@ -39,34 +39,34 @@
 	date = {2022},
 }
 
-@software{scripto,
+@misc{scripto,
 	title = {Scripto},
 	url = {https://scripto.org},
 }
 
-@software{tesseract,
+@misc{tesseract,
 	title = {Tesseract},
 	url = {https://tesseract-ocr.github.io},
 }
 
-@software{zooniverse,
+@misc{zooniverse,
 	title = {Zooniverse Project Builder},
 	url = {https://www.zooniverse.org/lab},
 }
 
-@software{abbyy,
+@misc{abbyy,
 	title = {{ABBYY} {FineReader}},
 	url = {https://pdf.abbyy.com},
 }
 
-@software{death,
+@misc{death,
 	title = {Death by Numbers},
 	url = {https://deathbynumbers.org/},
 	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
 	author = {Death by Numbers Project Team}
 }
 
-@software{mapping,
+@misc{mapping,
 	title = {Mapping Religious Ecologies},
 	url = {https://religiousecologies.org},
 	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
