### 2023-12-07 a0840f3
adding bibliography call
