### 2023-12-07 0192ab0
adding citations so references display right
