### 2023-12-07 bcf0328
tidy up in-line software and website references
