+++ paper.md
@@ -33,9 +33,9 @@ DataScribe is a structured data transcription module that extends the functional
 
 # Statement of Need
 
-Scholars often collect sources, such as government forms or institutional records, intending to transcribe them into datasets which can be analyzed or visualized. Many transcription programs such as ABBYY FineReader [@abbyy], Scripto for Omeka S [@scripto], Tesseract [@tesseract], and the Zooniverse Project Builder [@zooniverse] enable the manual or automated transcription into free-form text, but not into tables of data. The DataScribe module enables scholars to manually transcribe documents directly into a structured data format. Once scholars identify the structure of the data within their sources--such as numbers, dates, or controlled vocabularies--they can create forms that constrain and verify transcriptions done in the DataScribe interface. The transcriptions are then exported in tables of clean and tidy data that can be computationally analyzed or imported into a variety of analytical software programs. Because the module builds on Omeka S, scholars can also display transcriptions alongside the source images and metadata, crowdsource transcriptions, and publish their results on the web.
+Scholars often collect sources, such as government forms or institutional records, intending to transcribe them into datasets which can be analyzed or visualized. Many transcription programs such as @abbyy, @scripto for Omeka S, @tesseract, and @zooniverse enable the manual or automated transcription into free-form text, but not into tables of data. The DataScribe module enables scholars to manually transcribe documents directly into a structured data format. Once scholars identify the structure of the data within their sources--such as numbers, dates, or controlled vocabularies--they can create forms that constrain and verify transcriptions done in the DataScribe interface. The transcriptions are then exported in tables of clean and tidy data that can be computationally analyzed or imported into a variety of analytical software programs. Because the module builds on Omeka S, scholars can also display transcriptions alongside the source images and metadata, crowdsource transcriptions, and publish their results on the web.
 
-Projects using DataScribe include Death by Numbers (2016-ongoing) [@death], which is transcribing the seventeenth- and eighteenth-century London Bills of Mortality, and Mapping Religious Ecologies (2018-ongoing) [@mapping], which is transcribing the the 1926 United States Census of Religious Bodies. As part of the development of the module, the project team also created case study documentation for how DataScribe might be used to transcribe the London Bills of Mortality [@adasme:2022c], documentation on a 1903 plague outbreak in Chile in both Spanish and English [@adasme:2022a; @adasme:2022b], the 1926 United States Census of Religious Bodies [@swain:2022], and the 1950 United States Census [@brett:2022].
+Projects using DataScribe include @death (2016-ongoing), which is transcribing the seventeenth- and eighteenth-century London Bills of Mortality, and @mapping (2018-ongoing), which is transcribing the the 1926 United States Census of Religious Bodies. As part of the development of the module, the project team also created case study documentation for how DataScribe might be used to transcribe the London Bills of Mortality [@adasme:2022c], documentation on a 1903 plague outbreak in Chile in both Spanish and English [@adasme:2022a; @adasme:2022b], the 1926 United States Census of Religious Bodies [@swain:2022], and the 1950 United States Census [@brett:2022].
 
 # Acknowledgements
 
