### 2023-12-07 43667fb
reformatting website citations
bib converter was having trouble dealing with the "Mapping Religious Ecologies Project Team" and "Death by Numbers Project Team" group authorship references, which are duplicative of the project title so I am trying to see what it's like if I remove them
