+++ paper.bib
@@ -63,12 +63,10 @@
 	title = {Death by Numbers},
 	url = {https://deathbynumbers.org/},
 	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
-	author = {Death by Numbers Project Team}
 }
 
 @misc{mapping,
 	title = {Mapping Religious Ecologies},
 	url = {https://religiousecologies.org},
 	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
-	author = {Mapping Religious Ecologies Project Team}
 }
