+++ paper.bib
@@ -63,10 +63,12 @@
 	title = {Death by Numbers},
 	url = {https://deathbynumbers.org/},
 	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
+	date = {2016/2024}
 }
 
 @misc{mapping,
 	title = {Mapping Religious Ecologies},
 	url = {https://religiousecologies.org},
 	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
+	date = {2018/2024}
 }
