### 2023-12-07 2db9ecf
removing manual date references in paper.md
