+++ paper.bib
@@ -70,5 +70,5 @@
 	title = {Mapping Religious Ecologies},
 	url = {https://religiousecologies.org},
 	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
-	date = {2018/2024}
+	date = {2018}
 }
