### 2023-12-08 #5661 @luxaritas
I'm noticing that a lot of the references are showing up with just a date or even multiple non-disambiguated instances of `n.d.`. Looking at prior JOSS papers, it is very rare that no author is listed on a reference, even for websites (in which case you'd typically use some variation on the copyright holder, site owner, responsible organization, etc). Also at least for Tesseract, it may be preferable to use one of the citations listed in https://github.com/tesseract-ocr/tesseract/blob/main/CITATIONS.bib?

I don't know what JOSS preferences/standards on this are
