### 2023-12-08 050ec00
update Tesseract bib entry
experiment to see if I can force the PDF generator to create a proper "et al" citation, also adding a date for latest major release
