+++ paper.bib
@@ -47,6 +47,8 @@
 @misc{tesseract,
 	title = {Tesseract},
 	url = {https://tesseract-ocr.github.io},
+	author = {Weil, Stefan and Smith, Ray, and Podobny, Zdenko and Abdulkader, Ahmad and Antonova, Rika and Beato, Nicholas and Breidenbach, Jeff and Charron, Samuel and Cheatle, Phil and Crouch, Simon and Eger, David and Huddleston, Sheelagh and Johnson, Dan and Katikam, Rajesh and Kielbus, Thomas and Lee, Dar-Shyang and Liu, Zongyi and Moss, Robert and Newton, Chris and Reimer, Michael and Renn, Marius and Romano, Raquel and Russon, Christy and Saxena, Shobhit and Seaman, Mark and Shafait, Faisal and Takenaka, Hiroshi and Unnikrishnan, Ranjith and Wanke, Joern and Xiu, Ping Ping and Ziem, Andrew and Zuniga, Oscar and Regan, Jim and Barlow, James R. and Brechtken, Stefan and Breuel, Thomas and Dovev, Amit and Ettl, Martin and Devi Kumar, Shree and Metzger, Noah and Morris, Tom and Müller, Tobias and Pugin, Egor and Sachunsky, Robert and Schietekat, Raf and Vaidya, Sundar M. and Watts, Robin and White, Nick and Zaitsev, Alexander},
+	date = {2021},
 }
 
 @misc{zooniverse,
