+++ paper.bib
@@ -2,7 +2,7 @@
 @misc{adasme:2022a,
 	title = {Peste Bubónica en Iquique, 1903},
 	url = {https://datascribe.tech/casestudies/PesteBubonica_Iquique1903.pdf},
-	titleaddon = {{DataScribe}.tech},
+	publisher = {{DataScribe}.tech},
 	author = {Adasme, Hernán},
 	date = {2022},
 }
@@ -10,7 +10,7 @@
 @misc{adasme:2022b,
 	title = {Plague in Iquique, 1903},
 	url = {https://datascribe.tech/casestudies/Plague_Iquique1903.pdf},
-	titleaddon = {{DataScribe}.tech},
+	publisher = {{DataScribe}.tech},
 	author = {Adasme, Hernán},
 	date = {2022},
 }
@@ -18,7 +18,7 @@
 @misc{adasme:2022c,
 	title = {Death by Numbers},
 	url = {https://datascribe.tech/casestudies/DataScribe_BillsOfMortality_CaseStudy.pdf},
-	titleaddon = {{DataScribe}.tech},
+	publisher = {{DataScribe}.tech},
 	author = {Adasme, Hernán and Howlett, Daniel and Meyers, Emily},
 	date = {2022},
 }
@@ -26,7 +26,7 @@
 @misc{brett:2022,
 	title = {1950 United States Census},
 	url = {https://datascribe.tech/casestudies/DataScribeCaseStudy_1950CensusUS.pdf},
-	titleaddon = {{DataScribe}.tech},
+	publisher = {{DataScribe}.tech},
 	author = {Brett, Megan},
 	date = {2022},
 }
@@ -34,7 +34,7 @@
 @misc{swain:2022,
 	title = {Religious Ecologies},
 	url = {https://datascribe.tech/casestudies/CaseStudy_AmericanReligiousEcologies.pdf},
-	titleaddon = {{DataScribe}.tech},
+	publisher = {{DataScribe}.tech},
 	author = {Swain, Greta},
 	date = {2022},
 }
@@ -42,6 +42,8 @@
 @misc{scripto,
 	title = {Scripto},
 	url = {https://scripto.org},
+	author = {Hamner, Christopher and Safley, Jim and Nguyen, Kim and Brett, Megan and Leon, Sharon and Fahringer, Alyssa and Halabuk, James and Dauterive, Jessica and Albers, Ken and Ghajar, Lee Ann and Brennan, Sheila}
+	date = {2010/2019}
 }
 
 @misc{tesseract,
@@ -54,11 +56,15 @@
 @misc{zooniverse,
 	title = {Zooniverse Project Builder},
 	url = {https://www.zooniverse.org/lab},
+	author = {{Zooniverse Project Team}},
+	date = {2009},
 }
 
 @misc{abbyy,
 	title = {{ABBYY} {FineReader}},
 	url = {https://pdf.abbyy.com},
+	author = {ABBYY},
+	date = {2022},
 }
 
 @misc{death,
