+++ paper.bib
@@ -42,7 +42,7 @@
 @misc{scripto,
 	title = {Scripto},
 	url = {https://scripto.org},
-	author = {Hamner, Christopher and Safley, Jim and Nguyen, Kim and Brett, Megan and Leon, Sharon and Fahringer, Alyssa and Halabuk, James and Dauterive, Jessica and Albers, Ken and Ghajar, Lee Ann and Brennan, Sheila}
+	author = {Hamner, Christopher and Safley, Jim and Nguyen, Kim and Brett, Megan and Leon, Sharon and Fahringer, Alyssa and Halabuk, James and Dauterive, Jessica and Albers, Ken and Ghajar, Lee Ann and Brennan, Sheila},
 	date = {2010/2019}
 }
 
