### 2023-12-08 fb9a4a0
adding corporate authorships to bib
