+++ paper.bib
@@ -43,6 +43,7 @@
 	title = {Scripto},
 	url = {https://scripto.org},
 	author = {Hamner, Christopher and Safley, Jim and Nguyen, Kim and Brett, Megan and Leon, Sharon and Fahringer, Alyssa and Halabuk, James and Dauterive, Jessica and Albers, Ken and Ghajar, Lee Ann and Brennan, Sheila},
+	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
 	date = {2010/2019}
 }
 
@@ -55,8 +56,8 @@
 
 @misc{zooniverse,
 	title = {Zooniverse Project Builder},
-	url = {https://www.zooniverse.org/lab},
-	author = {{Zooniverse Project Team}},
+	url = {https://www.zooniverse.org},
+	author = {Johnson, Cliff and Chambers, Cory and Clement, Delilah and Trouille, Laura and Bouslog, Mark and Yuen, Michelle and Blickhan, Samantha and Miller, Sean and Wolfenbarger, Zach and Lintott, Chris and Noordin, Shaun A. and Roberts, Hayley and Mantha, Kameswara and Fortson, Lucy and Huebner, Sarah and Kiefer, Travis and Simmons, Brooke and Krawczyk, Coleman and Spiers, Helen and McMaster, Adam and Bowyer, Alex and Weiksnar, Alex and Swanson, Ali and Boyer, Amy and Simenstad, Andrea and Smith, Arfon and Rother, Becky and Newing, Beverley and Carstensen, Brian and Campbell, Adam and Troy, Chelsea and Snyder, Chris and Doogue, Christopher and McRoy, Darren and Wright, Darryl and Ostlund, Erik and Miller, Grant and Hines, Greg and Sewell, Hannah and van Singel, Heath and Dickinson, Hugh and Artz, James and Gupta, Jen and O'Donnell, Jim and Martin, Jordan and Feldt, Julie and Masters, Karen and Glover, Keanu and Borden, Kelly and Willett, Kyle and Whyte, Laura and Saeedi, Mariam and Veldthuis, Marten and Westwood, Mary and Bashraheel, Maysa and Schwamb, Meg and Beck, Melanie and Parrish, Michael and Simon, Molly and Ciemniak, Nicole and Malmed, Noah and Roper, Perry and Sankar, Ramana and Smethurst, Rebecca and Simpson, Robert and Schaaf, Robin and Hutchings, Roger and Costello, Rona and Aroney, Samuel and Kruk, Sandor and Allen, Sarah and Ishikawa, Sascha and Duca, Simone and Raden, Steve and Lynn, Stuart and Maidel, Veronica and Van Hyning, Victoria and Granger, Will},
 	date = {2009},
 }
 
@@ -70,6 +71,7 @@
 @misc{death,
 	title = {Death by Numbers},
 	url = {https://deathbynumbers.org/},
+	author = {{Death by Numbers Project Team}},
 	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
 	date = {2016/2024}
 }
@@ -77,6 +79,7 @@
 @misc{mapping,
 	title = {Mapping Religious Ecologies},
 	url = {https://religiousecologies.org},
+	author = {{Mapping Religious Ecologies Project Team}},
 	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
-	date = {2018}
+	date = {2018/2026}
 }
