### 2023-12-08 e8bfbbb
updating paper to match new citations
