### 2023-12-08 9456597
removing team lines from bib
