+++ paper.bib
@@ -71,7 +71,6 @@
 @misc{death,
 	title = {Death by Numbers},
 	url = {https://deathbynumbers.org/},
-	author = {{Death by Numbers Project Team}},
 	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
 	date = {2016/2024}
 }
@@ -79,7 +78,6 @@
 @misc{mapping,
 	title = {Mapping Religious Ecologies},
 	url = {https://religiousecologies.org},
-	author = {{Mapping Religious Ecologies Project Team}},
 	publisher = {Roy Rosenzweig Center for History and New Media, George Mason University},
 	date = {2018/2026}
 }
