+++ paper.bib
@@ -1,6 +1,6 @@
 
 @misc{adasme:2022a,
-	title = {Peste Bubónica en Iquique, 1903},
+	title = {Peste Bubónica en Iquique, 1903: Transcripción de datos no tabulares usando DataScribe},
 	url = {https://datascribe.tech/casestudies/PesteBubonica_Iquique1903.pdf},
 	publisher = {{DataScribe}.tech},
 	author = {Adasme, Hernán},
@@ -8,7 +8,7 @@
 }
 
 @misc{adasme:2022b,
-	title = {Plague in Iquique, 1903},
+	title = {Plague in Iquique, 1903: Transcribing non tabular data using DataScribe},
 	url = {https://datascribe.tech/casestudies/Plague_Iquique1903.pdf},
 	publisher = {{DataScribe}.tech},
 	author = {Adasme, Hernán},
@@ -24,7 +24,7 @@
 }
 
 @misc{brett:2022,
-	title = {1950 United States Census},
+	title = {An Exercise in Iteration: Transcribing the 1950 United States Census with DataScribe},
 	url = {https://datascribe.tech/casestudies/DataScribeCaseStudy_1950CensusUS.pdf},
 	publisher = {{DataScribe}.tech},
 	author = {Brett, Megan},
@@ -32,7 +32,7 @@
 }
 
 @misc{swain:2022,
-	title = {Religious Ecologies},
+	title = {American Religious Ecologies},
 	url = {https://datascribe.tech/casestudies/CaseStudy_AmericanReligiousEcologies.pdf},
 	publisher = {{DataScribe}.tech},
 	author = {Swain, Greta},
