### 2024-01-06 #5661 @arfon
@jmotis – I don't think the BibTeX entries with multiple year values are valid (e.g., https://github.com/chnm/Datascribe-module/blob/JOSS/paper.bib#L82). Please could you correct those three entries to only have a single year value?
