### 2024-01-07 fd10550
remove date ranges from bib
