+++ paper.bib
@@ -44,7 +44,7 @@
 	url = {https://scripto.org},
 	author = {Hamner, Christopher and Safley, Jim and Nguyen, Kim and Brett, Megan and Leon, Sharon and Fahringer, Alyssa and Halabuk, James and Dauterive, Jessica and Albers, Ken and Ghajar, Lee Ann and Brennan, Sheila},
 	publisher = {{Roy Rosenzweig Center for History and New Media, George Mason University}},
-	date = {2010/2019}
+	date = {2010}
 }
 
 @misc{tesseract,
@@ -72,12 +72,12 @@
 	title = {Death by Numbers},
 	url = {https://deathbynumbers.org/},
 	publisher = {{Roy Rosenzweig Center for History and New Media, George Mason University}},
-	date = {2016/2024}
+	date = {2016}
 }
 
 @misc{mapping,
 	title = {Mapping Religious Ecologies},
 	url = {https://religiousecologies.org},
 	publisher = {{Roy Rosenzweig Center for History and New Media, George Mason University}},
-	date = {2018/2026}
+	date = {2018}
 }
